+++ packages/docstage/src/docstage/app_keys.py
@@ -2,12 +2,12 @@
 
 from aiohttp import web
 
-from docstage.core.cache import FileCache
+from docstage.core.cache import PageCache
 from docstage.core.renderer import PageRenderer
 from docstage.core.site import SiteLoader
 
 renderer_key = web.AppKey("renderer", PageRenderer)
 site_loader_key = web.AppKey("site_loader", SiteLoader)
-cache_key = web.AppKey("cache", FileCache)
+cache_key = web.AppKey("cache", PageCache)
 verbose_key = web.AppKey("verbose", bool)
 live_reload_enabled_key = web.AppKey("live_reload_enabled", bool)